// src/modules/reports/models/report_type.model.ts
export enum ReportType {
    'func',
    'loading',
    'save-data',
    'unreadable',
    'conflict',
    'no-follow',
    'connection',
    'flow',
    'format',
    'freeze',
    'other',
}

export enum ReportTypeKeys {
    FUNC = 'func',
    LOADING = 'loading',
    SAVE = 'save-data',
    UNREADABLE = 'unreadable',
    CONFLICT = 'conflict',
    NO_FOLLOW = 'no-follow',
    CONN = 'connection',
    FLOW = 'flow',
    FORMAT = 'format',
    FREEZE = 'freeze',
    OTHER = 'other',
}